fix(Table of Contents): :bug: Wait for all async calls to return (properly).
Previously, once the for loop was finished the page was considered to be "loaded" (which of course it was, it's all very lightweight). But the proper thing to do would be to keep a running count of all the async calls, since they may return out of order, and check after each one whether we've added up to the total of the for loop.